move query parsing test
